requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.4.0
scipy>=1.6.0
pywavelets>=1.1.0
tensorflow>=2.4.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
PyQt5>=5.15.0
cryptography>=3.4.0 